requirements.txt
discord.py-self==2.0.0
audioop-lts>=0.2.2
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
Pillow>=10.4.0
qrcode==7.4.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml>=5.3.0
aiohttp>=3.11.0
pyttsx3==2.90
pydub==0.25.1
ffmpeg-python==0.2.0
